+++ R/plot2.R
@@ -26,8 +26,8 @@
 #'   following values are possible, for details, see plot: "p" for points, "l"
 #'   for lines, "b" for both points and lines, "c" for empty points joined by
 #'   lines, "o" for overplotted points and lines, "s" and "S" for stair steps
-#'   and "h" for histogram-like vertical lines. Finally, "n" does not produce
-#'   any points or lines.
+#'   and "h" for histogram-like vertical lines. "n" does not produce
+#'   any points or lines. "pointrange" draws point-range plots.
 #' @param xlim the x limits (x1, x2) of the plot. Note that x1 > x2 is allowed
 #'   and leads to a ‘reversed axis’. The default value, NULL, indicates that
 #'   the range of the `finite` values to be plotted should be used.
@@ -94,7 +94,7 @@
 #'   be calling `dev.off()` to reset all `par` settings to their defaults.)
 #' @param subset,na.action,drop.unused.levels arguments passed to `model.frame`
 #'   when extracting the data from `formula` and `data`.
-#' @param ymin,ymax minimum and maximum coordinates of the point range. Only used when `type="pr"`.
+#' @param ymin,ymax minimum and maximum coordinates of the point range. Only used when `type="pointrange"`.
 #' @param ... 	other `graphical` parameters (see `par` and also the "Details"
 #'   section of `plot`).
 #'   
@@ -259,8 +259,8 @@ plot2.default = function(
   if (is.null(xlim)) xlim = range(x, na.rm = TRUE)
   if (is.null(ylim)) ylim = range(y, na.rm = TRUE)
 
-  if (!is.null(ymin)) ylim[1] <- min(c(ylim, ymin))
-  if (!is.null(ymax)) ylim[2] <- max(c(ylim, ymax))
+  if (!is.null(ymin)) ylim[1] = min(c(ylim, ymin))
+  if (!is.null(ymax)) ylim[2] = max(c(ylim, ymax))
 
   if (!is.null(by)) {
     l = list(x=x, y=y)
@@ -408,7 +408,22 @@ plot2.default = function(
   if (!is.null(grid)) grid
   
   # draw the points/lines
-  if (type %in% "p") {
+  if (type %in% "pointrange") { # segments before point
+    invisible(
+      lapply(
+        seq_along(split_data),
+        function(i) {
+          graphics::segments(
+            x0 = seq_along(split_data[[i]]$x),
+            y0 = split_data[[i]]$ymin,
+            x1 = seq_along(split_data[[i]]$x),
+            y1 = split_data[[i]]$ymax
+          )
+        }
+      )
+    )
+  } 
+  if (type %in% c("p", "pointrange")) {
     invisible(
       lapply(
         seq_along(split_data),
@@ -434,28 +449,6 @@ plot2.default = function(
             y = split_data[[i]]$y,
             col = col[i],
             type = type,
-            pch = pch[i],
-            lty = lty[i]
-          )
-        }
-      )
-    )
-  } else if (type %in% "pr") {
-    invisible(
-      lapply(
-        seq_along(split_data),
-        function(i) {
-          segments(
-            x0 = seq_along(split_data[[i]]$x),
-            y0 = split_data[[i]]$ymin,
-            x1 = seq_along(split_data[[i]]$x),
-            y1 = split_data[[i]]$ymax
-          )
-          points(
-            x = split_data[[i]]$x,
-            y = split_data[[i]]$y,
-            col = col[i],
-            type = "p",,
             pch = pch[i],
             lty = lty[i]
           )